requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
python-dotenv==1.0.0
supabase==1.0.3
pydantic==2.5.2
httpx==0.23.3